Add equal magic method for infer types
