+++ python/kserve/test/test_infer_type.py
@@ -135,7 +135,7 @@ class TestInferRequest:
                                 from_grpc=True
                                 )
         res = InferRequest.from_grpc(infer_req)
-        assert InferRequestMatcher(expected).__eq__(res) is True
+        assert res == expected
 
     class TestInferResponse:
         def test_to_rest(self):
@@ -255,90 +255,4 @@ class TestInferRequest:
                                      from_grpc=True
                                      )
             res = InferResponse.from_grpc(infer_res)
-            assert InferResponseMatcher(expected).__eq__(res) is True
-
-
-class InferInputMatcher:
-    def __init__(self, infer_input: InferInput):
-        self.input = infer_input
-
-    def __eq__(self, other):
-        if not isinstance(other, InferInput):
-            return False
-        if self.input.name != other.name:
-            return False
-        if self.input.shape != other.shape:
-            return False
-        if self.input.datatype != other.datatype:
-            return False
-        if self.input.parameters != other.parameters:
-            return False
-        if self.input.data != other.data:
-            return False
-        return True
-
-
-class InferRequestMatcher:
-    def __init__(self, req: InferRequest):
-        self.req = req
-
-    def __eq__(self, other):
-        if not isinstance(other, InferRequest):
-            return False
-        if self.req.model_name != other.model_name:
-            return False
-        if self.req.id != other.id:
-            return False
-        if self.req.from_grpc != other.from_grpc:
-            return False
-        if self.req.parameters != other.parameters:
-            return False
-        if len(self.req.inputs) != len(other.inputs):
-            return False
-        for i in range(len(self.req.inputs)):
-            if not InferInputMatcher(self.req.inputs[i]).__eq__(other.inputs[i]):
-                return False
-        return True
-
-
-class InferOutputMatcher:
-    def __init__(self, infer_input: InferOutput):
-        self.output = infer_input
-
-    def __eq__(self, other):
-        if not isinstance(other, InferOutput):
-            return False
-        if self.output.name != other.name:
-            return False
-        if self.output.shape != other.shape:
-            return False
-        if self.output.datatype != other.datatype:
-            return False
-        if self.output.parameters != other.parameters:
-            return False
-        if self.output.data != other.data:
-            return False
-        return True
-
-
-class InferResponseMatcher:
-    def __init__(self, res: InferResponse):
-        self.res = res
-
-    def __eq__(self, other):
-        if not isinstance(other, InferResponse):
-            return False
-        if self.res.model_name != other.model_name:
-            return False
-        if self.res.id != other.id:
-            return False
-        if self.res.from_grpc != other.from_grpc:
-            return False
-        if self.res.parameters != other.parameters:
-            return False
-        if len(self.res.outputs) != len(other.outputs):
-            return False
-        for i in range(len(self.res.outputs)):
-            if not InferOutputMatcher(self.res.outputs[i]).__eq__(other.outputs[i]):
-                return False
-        return True
+            assert res == expected
